refactor(Rust): Implement `kernels` crate
