+++ rust/kernel-jupyter/Cargo.toml
@@ -8,7 +8,7 @@ version = "0.0.0"
 default = ["zmq/vendored"]
 
 [dependencies]
-chrono = "0.4.19"
+chrono = {version = "0.4.19", features = ["serde"]}
 codec-ipynb = {path = "../codec-ipynb", version = "0.0.0"}
 defaults = "0.2.0"
 dirs = "4.0.0"
@@ -25,11 +25,10 @@ serde = "1.0.130"
 serde_json = "1.0.68"
 serde_with = "1.11.0"
 sha2 = "0.9.8"
-tokio = "1.13.0"
+tokio = {version = "1.12.0", features = ["full"]}
 tracing = "0.1.29"
 uuid-utils = {path = "../uuid-utils", version = "0.0.0"}
 zmq = {version = "0.9.2", default-features = false}
 
 [dev-dependencies]
 test-utils = {path = "../test-utils", version = "0.0.0"}
-tokio = {version = "1.12.0", features = ["full"]}
