+++ rust/stencila/Cargo.toml
@@ -38,8 +38,6 @@ plugins-link = ["plugins"]
 plugins-python = ["plugins"]
 plugins-r = ["plugins"]
 
-kernels-jupyter = ["zmq/vendored"]
-
 upgrade = ["self_update"]
 
 import-elife = ["minidom"]
@@ -86,7 +84,6 @@ default = [
   "plugins-javascript",
   "plugins-python",
   "plugins-r",
-  "kernels-jupyter",
   "upgrade",
   "import-elife",
   "import-github",
@@ -116,7 +113,6 @@ gitignore = "=1.0.7"
 glob = "=0.3.0"
 handlebars = "=4.1.3"
 hex = "0.4.3"
-hmac = "0.11.0"
 humantime = "=2.1.0"
 ignore = "=0.4.18"
 Inflector = "0.11.4"
@@ -128,7 +124,6 @@ jsonwebtoken = { version = "=7.2.0", optional = true }
 maplit = "=1.0.2"
 mime_guess = "=2.0.3"
 minidom = { version = "=0.13.0", optional = true }
-nanoid = "=0.4.0"
 notify = "=4.0.17"
 once_cell = "=1.8.0"
 path-slash = "=0.1.4"
@@ -176,7 +171,6 @@ urlencoding = "2.1.0"
 validator = { version = "=0.14.0", features = ["derive"] }
 warp = { version = "=0.3.1", optional = true }
 webbrowser = { version = "=0.5.5", optional = true }
-zmq = { version = "=0.9.2", optional = true, default-features = false }
 
 binaries = { path = "../binaries", version = "0.0.0", optional = true }
 cli-utils = { path = "../cli-utils", version = "0.0.0", optional = true }
